Adding tlslite support.
tlslite is a pure python implimentation of TLS. It supports a bunch of libraries
to make the library much faster (OpenSSL, pyCrypto, pyGMP, etc)

For some reason when connecting to my local tpserver-cpp the following patch
needs to be applied to tlslite.

Lines around 73 need to be change to look like the below.
 73         if not list(rsaOID) in (
 74                 [6, 9, 42, 134, 72, 134, 247, 13, 1, 1, 1, 5, 0],
 75                 [6, 9, 42, 134, 72, 134, 247, 13, 1, 1, 1]):
 76             raise SyntaxError("Unrecognized AlgorithmIdentifier")
.